+++ test/src/factories/placemark_factory.dart
@@ -5,17 +5,16 @@ import 'position_factory.dart';
 class PlacemarkFactory {
   static Placemark createMockPlacemark() {
     return Placemark(
-      administrativeArea: 'Overijssel',
-      country: 'Netherlands',
-      isoCountryCode: 'NL',
-      locality: 'Enschede',
-      name: 'Gronausestraat',
-      position: PositionFactory.createMockPosition(),
-      postalCode: '',
-      subAdministrativeArea: 'Enschede',
-      subLocality: 'Enschmarke',
-      subThoroughfare: '',
-      thoroughfare: 'Gronausestraat'
-    );
+        administrativeArea: 'Overijssel',
+        country: 'Netherlands',
+        isoCountryCode: 'NL',
+        locality: 'Enschede',
+        name: 'Gronausestraat',
+        position: PositionFactory.createMockPosition(),
+        postalCode: '',
+        subAdministrativeArea: 'Enschede',
+        subLocality: 'Enschmarke',
+        subThoroughfare: '',
+        thoroughfare: 'Gronausestraat');
   }
 }